cast axis and angle to float

+++ airo-spatial-algebra/airo_spatial_algebra/se3.py
@@ -113,11 +113,12 @@ class SE3Container:
     @property
     def orientation_as_axis_angle(self) -> AxisAngleType:
         angle, axis = self.se3.angvec()
-        return axis, angle
+        return axis.astype(np.float64), float(angle)
 
     @property
     def orientation_as_rotation_vector(self) -> Vector3DType:
         axis, angle = self.orientation_as_axis_angle
+
         if axis is None:
             return np.zeros(3)
         return angle * axis
